requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1

SQLAlchemy==2.0.29
psycopg2-binary==2.9.9

python-jose[cryptography]==3.3.0
passlib==1.7.4
python-dotenv==1.0.1
python-multipart==0.0.9

pytest==8.3.3